tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/139.html

<!DOCTYPE html>
<html><head>
	<title>scheduler: SVG script nested external in inline</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<div id="log"></div>
<script>
  var t = async_test(undefined, {timeout:3000});
</script>
<svg>
<script>
t.step(function() {
log("inline script #1");
var a = {
  <script xlink:href="scripts/include-1.js">
    t.step(function() {assert_unreached()})
  </script>
a:1}
log("end inline script #1");
});
</script>
</svg>
<script>
onload = t.step_func(function() {
  assert_array_equals(eventOrder, ["external script #1", "inline script #1", "end inline script #1"]);
  t.done();
});
</script>
